python_generic/requirements.txt
pandas                 == 2.2.2            
numpy                  == 1.26.4
pip                    == 22.3.1
python-dateutil        == 2.9.0.post0
pytz                   == 2024.1
setuptools             == 65.5.1
six                    == 1.16.0
tzdata                 == 2024.1
wheel                  == 0.40.0
contourpy              == 1.2.1
cycler                 == 0.12.1
certifi                == 2024.6.2
matplotlib             == 3.9.0
packaging              == 24.1
pillow                 == 10.3.0
pyparsing              == 3.1.2
python-dateutil        == 2.9.0.post0
mysql-connector-python == 9.0.0
Django                 == 5.1.1
